\name{lpoly_make_system}
\alias{lpoly_make_system}
\title{LPoly system constructor
Returns and XPtr to an lpoly_system_type object}
\usage{
lpoly_make_system(cm, trm)
}
\arguments{
  \item{cm}{Model coefficient matrix}

  \item{trm}{Model term specification, as generated by
  lpoly_model_spec (each row represents the powers of the
  different factors in a term)}
}
\description{
LPoly system constructor Returns and XPtr to an
lpoly_system_type object
}

