\name{lpoly_make_system_xptr}
\alias{lpoly_make_system_xptr}
\title{LPoly system constructor}
\usage{
lpoly_make_system_xptr(cm, trm)
}
\arguments{
  \item{cm}{Model coefficient matrix}

  \item{trm}{Model term specification, as generated by
  lpoly_model_spec (each row represents the powers of the
  different factors in a term)}
}
\description{
Returns an XPtr to an lpoly_system_type object
}

