\name{lpoly_model_matrix}
\alias{lpoly_model_matrix}
\title{LPoly model matrix generator
Builds the model matrix from the data using the lpoly_system_type object}
\usage{
lpoly_model_matrix(lps, data)
}
\arguments{
  \item{lps}{XPtr to an lpoly_system_type object created by
  a call to lpoly_make_system}

  \item{data}{One column per measurement variable}
}
\description{
LPoly model matrix generator Builds the model matrix from
the data using the lpoly_system_type object
}

