\docType{package}
\name{SimpleSDESampler}
\alias{SimpleSDESampler}
\alias{SimpleSDESampler-package}
\title{SimpleSDESampler.}
\description{
Runs simulations on SDE systems
}
\section{Included functions}{
  \itemize{ \item \code{\link{synthetic.dataset}} \item
  \code{\link{synthetic.dataset.quick}} \item
  \code{\link{lpoly_implicit_sde}} \item
  \code{\link{lpoly_implicit_sde_averages}} }
}

